feat: add task board and patch status
