+++ AGENTS.md
@@ -11,6 +11,7 @@
 - Текстовые сообщения бота в `bot/src/messages.ts` должны быть на русском.
 - Контексты React держите в отдельных файлах, провайдеры экспортируйте из этих файлов.
 - Логирование сохраняет только изменения задач, профиля и вход; обращения к `/api/v1/logs` не записываются.
+- Обновление статуса задачи выполняется PATCH `/api/v1/tasks/:id/status`.
 
 ## Качество и тесты
