+++ README.md
@@ -10,6 +10,7 @@
 - Веб‑панель администратора на базе TailAdmin.
 - REST API с документацией Swagger по пути `/api-docs`.
 - Поддержка расчёта маршрутов через сервис OSRM.
+- Канбан-доска задач с перетаскиванием и обновлением статуса.
 - Запросы к OSRM кешируются в памяти или Redis на 10 минут и очищаются при
   изменении задач, матрица `/table` ограничена по размеру и частоте
   переменными `ROUTE_TABLE_GUARD`, `ROUTE_TABLE_MAX_POINTS` и
