+++ ROADMAP.md
@@ -24,6 +24,7 @@
 - Dependabot еженедельно обновляет npm-зависимости, CI запускает `audit-ci` с порогом `high`.
 - Стабилизированы запросы OSRM: алгоритм MLD по умолчанию, кеш с
   инвалидированием при изменении задач и ограничение размера и частоты матриц.
+- Добавлена канбан-доска задач с переключением таблица/доска и перетаскиванием для изменения статуса.
 
 ## Ближайшие задачи
