+++ futureUpdates.md
@@ -7,7 +7,7 @@ GitHub
 Провести ревизию DI-контейнера: определить, какие классы/сервисы можно описать интерфейсами, и зарегистрировать их с именованными токенами. Удалить использование контейнера как сервис-локатора (где сервисы запрашиваются по string-ключам из произвольных мест) – вместо этого внедрять зависимости через конструкторы (благо tsyringe это позволяет). Это сделает структуру зависимостей явной.
 Удалить дублированный код и модули. В roadmap упомянуто удаление дублирующего модуля services/tasks.ts
 GitHub
- – проверить, что таких случаев больше нет (видимо, оставался старый файл после рефакторинга). Убедиться, что логика не продублирована между ботом и API: например, формирование текста задачи есть и в боте (для сообщения) и в API (для ответа) – может, стоит вынести общее в утилиты, чтобы избегать рассинхрона.
+– проверить, что таких случаев больше нет (видимо, оставался старый файл после рефакторинга). Убедиться, что логика не продублирована между ботом и API: например, формирование текста задачи есть и в боте (для сообщения) и в API (для ответа) – может, стоит вынести общее в утилиты, чтобы избегать рассинхрона.
 Улучшить обработку ошибок в API: уже внедрен общий error handler с форматом Problem Details
 GitHub
 . Можно расширить собственные классы ошибок (например, NotFoundError, ValidationError) и использовать их в сервисах, чтобы error handler мог возвращать более семантичные сообщения. Это упростит отладку и интеграцию (клиенту будет понятно, что за ошибка).
@@ -18,7 +18,7 @@ GitHub
 Обновить ROADMAP и technical_manual с учётом выполненного на этапе. Например, зафиксировать, что Command/Strategy в работе, какие рефакторинги сделаны.
 Результат этапа 1: кодовая база очищена от оставшихся технических долгов, все процессы автоматизации отлажены. Проект готов для безопасного добавления крупных новых фич без угрозы нарушить существующий функционал.
 Этап 2: UX-улучшения и базовый браузерный режим (ближайшие 1-2 месяца). На этом этапе фокус на видимых улучшениях для пользователей:
-Реализовать Kanban-доску для задач. Начать можно с представления только для администраторов (так как они основной пользующийся веб-интерфейсом контингент). Канбан можно внедрить параллельно со старой таблицей: дать пользователю переключатель “Таблица/Доска”. Карточка задачи будет показывать название, ответственного, статус, может быть срок. Перетаскивание меняет статус (и вызывает PATCH /tasks/:id/status). Нужно позаботиться о правах – обычному пользователю, скорее всего, нельзя менять статус, значит для них либо доска только чтение, либо недоступна. Впоследствии эту же доску (или упрощенную её версию) можно показать и в Telegram через WebApp (если влезет по ширине, хотя на мобильном сложнее). Но приоритет – браузер.
+Канбан-доска задач реализована: есть переключатель "Таблица/Доска", перетаскивание меняет статус через PATCH `/tasks/:id/status`. Остаётся оценить права доступа для обычных пользователей и возможность упрощённой версии в Telegram WebApp.
 Добавить календарный вид. Если данных о сроках мало, календарь может быть не первостепенным. Однако, можно реализовать хотя бы отображение дедлайнов. Например, вывести небольшой календарь сбоку, где числа отмечены, на какие даты есть задачи, и при клике – список задач за этот день. Более сложную реализацию (drag&drop задач на календарь для смены даты) можно запланировать на потом. Главное – дать возможность визуально видеть нагрузку по дням.
 Улучшить форму создания/редактирования задач с точки зрения UX: убедиться, что порядок полей логичен (в roadmap указано, что поле "Название" переместили вверх
 GitHub
@@ -39,10 +39,10 @@ Onboarding и подсказки. Разработать небольшие сп
 Расширяемость и интеграции. Открыть API для интеграций с другими системами или предоставить вебхуки. Например, по событию создания или изменения задачи отправлять POST-запрос на указанный URL (настраиваемый в Settings) – это позволит внешним системам реагировать (такое требуется, если, скажем, компания хочет связать наш таск-трекер со своей CRM). Аналогично, можно интегрировать входящие вебхуки – чтобы из внешней системы создавать задачи (нужна авторизация по токену). Это шаг в сторону платформенности, что увеличит ценность продукта в экосистеме.
 Дополнительные модули. Проанализировать, какие функции предлагают конкуренты (например, модуль аналитики: отчёты по выполненным задачам, эффективность исполнителей). У нас уже есть KPI-эндпоинт /api/v1/tasks/report/summary – можно на его основе в веб-интерфейсе сделать раздел "Отчёты" с графиками: сколько задач создано/выполнено за период, среднее время выполнения, и т.п. Также упоминалось расширение модели задачи для логистики, закупок
 GitHub
- – возможно, предстоит добавить новые поля или типы задач. Эти новые модули надо внедрять аккуратно, сохраняя общую стабильность системы.
+– возможно, предстоит добавить новые поля или типы задач. Эти новые модули надо внедрять аккуратно, сохраняя общую стабильность системы.
 Оптимизация маршрутов и логистика. Если задача связана с маршрутами, можно внедрить более сложные алгоритмы оптимизации (в roadmap отмечена оптимизация для 1–3 машин, режимы angle или trip
 GitHub
- уже реализованы). В будущем можно добавить, к примеру, расчет областей охвата, расчет времени прибытия и пр. Но это узкоспециализированные улучшения, их делать по мере требований пользователей.
+уже реализованы). В будущем можно добавить, к примеру, расчет областей охвата, расчет времени прибытия и пр. Но это узкоспециализированные улучшения, их делать по мере требований пользователей.
 Мобильное приложение? Хотя сейчас ставка на Telegram Mini App + Web, нельзя исключать, что в какой-то момент может понадобиться отдельное мобильное приложение (например, для работы вне Telegram, с офлайн-режимом). Поскольку у нас уже есть API и веб-клиент, технологии позволяют относительно легко создать React Native или Flutter приложение, используя те же API. Это не план непосредственного будущего, но стратегически – быть готовыми, что продукт может эволюционировать и в standalone приложение, если это потребуется рынку.
 Больше кастомизации для админа. Продолжить работу по предоставлению админу контроля: например, конструктор форм – крайне желанная фича во многих системах, когда админ может добавлять пользовательские поля к задачам. Реализовать это непросто (потребуется динамическая схема БД и UI), но постепенно можно идти к этому: сначала дать менять справочники (что будет сделано на этапе 2), потом позволить добавлять простые дополнительные поля (например, поле “Бюджет” типа число, поле “Комментарий клиента” тип текст). Хранить их можно в отдельной коллекции (типа TaskExtra), либо в JSON-поле задачи. UI будет генерировать дополнительные поля на основе конфигурации. Такой уровень гибкости превратит продукт из "фиксированного" в более "настраиваемый", что привлекает продвинутых пользователей.
 Усиление аналитики использования. Внедрить сбор анонимной статистики использования функций (с согласия, если open-source проект). Это поможет узнать, какие функции реально используются, а какие нет, и направлять усилия на то, что важно.
